Testing some features so im commiting first

+++ src/pages/Students/Students List/Students List.tsx
@@ -1,12 +1,13 @@
 import React, { useState, useEffect } from 'react';
 import { db } from '../../../firebase-config';
 import { collection, query, orderBy, getDocs, deleteDoc, doc } from 'firebase/firestore';
-import { FaMapMarkerAlt, MdDelete, MdModeEditOutline } from '../../../assets/IconImports';
+import { IconImports } from '../../../assets';
 import { useNavigate } from 'react-router-dom';
 
 type Student = {
   id: string;
   name: string;
+  question: string;
   bookOfStudy: string;
   pinnedLocation: {
     lat: number;
@@ -30,6 +31,7 @@ const StudentsList: React.FC = () => {
           id: doc.id,
           name: data.name,
           bookOfStudy: data.bookOfStudy,
+          question: data.question,
           pinnedLocation: data.pinnedLocation,
         };
       });
@@ -43,19 +45,26 @@ const StudentsList: React.FC = () => {
   const handleTraceLocation = (lat: number, lng: number) => {
     setTimeout(() => {
       window.open(`https://maps.google.com/?q=${lat},${lng}`, '_blank');
-    }, 1500); 
+    }, 1500);
 
   };
 
   const deleteStudent = async (studentId: string) => {
     try {
       const studentRef = doc(db, 'Students', studentId);
-      await deleteDoc(studentRef);
-      setStudents((prevStudents) => prevStudents.filter((student) => student.id !== studentId));
+      const result = confirm("Do you want to delete this student?");
+  
+      if (result) {
+        await deleteDoc(studentRef);
+        setStudents((prevStudents) => prevStudents.filter((student) => student.id !== studentId));
+        alert('Student Deleted Successfully!');
+      }
     } catch (error) {
       console.error('Error deleting student:', error);
+      alert('Student Could Not Be Deleted!');
     }
   };
+  
 
   const editStudent = (id: string) => {
     navigate(`/students/update-students/${id}`);
@@ -68,12 +77,13 @@ const StudentsList: React.FC = () => {
           <form>
             <h3>{student.name}</h3>
             <p>{student.bookOfStudy}</p>
+            <p>{student.question}</p>
             <div onClick={() => handleTraceLocation(student.pinnedLocation.lat, student.pinnedLocation.lng)}>
               Trace Location
-              <FaMapMarkerAlt className="icon-locate" />
+              <IconImports.FaMapMarkerAlt className="icon-locate" />
             </div>
-            <MdDelete className="icon-delete" onClick={() => deleteStudent(student.id)} />
-            <MdModeEditOutline className="icon-edit" onClick={() => editStudent(student.id)}/>
+            <IconImports.MdDelete className="icon-delete" onClick={() => deleteStudent(student.id)} />
+            <IconImports.MdModeEditOutline className="icon-edit" onClick={() => editStudent(student.id)} />
           </form>
         </div>
       ))}
